fix(symbolic): make the type-guard test real, pin composition order, split transform.rs
Two review findings on the type-check fix, both with the maintainer's own
probe evidence:

1. non_placement_entity_is_not_silently_misread used an
   IfcCartesianTransformationOperator3D fixture, whose attr 0 is an
   IfcDirection — #2355's pre-existing mandatory-Location check fires
   before the new type guard is ever reached, so deleting the guard left
   the test green. Swapped to an IfcAxis1Placement fixture (Location,
   Axis — attribute-count-compatible with IfcAxis2Placement2D, so it
   actually exercises the guard): mutation-verified RED (fabricated
   tx=5, ty=6, 90-degree rotation) with the guard deleted, GREEN restored.
   Also corrected the doc comment's now-inaccurate claim about its own
   fixture.

2. compose_transforms(&context_transform, &placement_transform)'s
   argument order was unpinned — every existing fixture uses an identity
   IfcLocalPlacement, so compose(a, I) == compose(I, a) regardless of
   order. Added composition_order_is_context_outer_placement_inner with
   a non-identity ObjectPlacement and a translated+rotated WCS;
   mutation-verified RED when the two arguments are flipped (yields
   [13, -24, 13, -25] instead of the correct [6, -23, 6, -24]).

3. The type-guard fix took transform.rs 399 -> 414, needing a new
   module_size_allowlist.txt row despite #2352 promising the next
   addition there would split the module instead. Moved
   parse_cartesian_transformation_operator (the largest, self-contained
   function) into a new sibling operator.rs, and moved the test module
   to a flat transform_tests.rs sibling via #[path], matching the
   step_text.rs precedent instead of the implicit transform/
   subdirectory. transform.rs is now 291 lines; the allowlist row is
   deleted and the ratchet passes without it.
